Revert "Implement UMS login test automation with enhanced click detection and comprehensive logging"
